Check if Feel Lonely skill present before adding to mood (#100)
Currently, if the player is not Emotionally Chipped, they don't have access to the skill Feel Lonely. The current script mood setup tries to cast Feel Lonely based only on the preference _feelLonelyUsed, and breaks if the skill is not present. This adds the check for it to run while not Emotionally Chipped.